+++ src-webviews/src/pages/components.ts
@@ -2,7 +2,6 @@ import { shallowRef } from 'vue';
 
 import Actions from './actions/Actions.vue';
 import Audio from './audio/Audio.vue';
-import Designs from './designs/Designs.vue';
 import Icons from './icons/Icons.vue';
 import InputBox from './input/InputBox.vue';
 import Job from './job/Job.vue';
@@ -14,7 +13,6 @@ export const CORE_IMPORTS = {
     Actions: shallowRef(Actions),
     Audio: shallowRef(Audio),
     BackgroundImage: shallowRef(BackgroundImageVue),
-    Designs: shallowRef(Designs),
     Icons: shallowRef(Icons),
     InputBox: shallowRef(InputBox),
     Job: shallowRef(Job),