+++ src/core/client/startup.ts
@@ -75,7 +75,6 @@ import './streamers/worldNotifications.js';
 import './streamers/worldNotificationsVirtual.js';
 
 // Systems Default
-import './systems/defaults/displayId.js';
 import './systems/defaults/time.js';
 import './systems/defaults/toolbar.js';